implement sign-out, fix coveralls
