+++ app/views/layout.erb
@@ -8,6 +8,12 @@
   <body>
     <%if current_user %>
       Welcome, <%= current_user.nom_name%>
+
+      <form method='post' action='/sessions'>
+        <input type='hidden' name='_method' value='delete'>
+        <input type='submit' value='Sign out'>
+      </form>
+
     <%end%>
 
     <% if flash[:errors] && !flash[:errors].empty? %>
@@ -17,6 +23,12 @@
           <li><%= error %></li>
         <% end %>
     <% end %>
+
+    <% if flash[:notice] %>
+        <div id='notice'>
+          <%= flash[:notice] %>
+        </div>
+      <% end %>
 
     <%= yield %>
   </body>
