+++ spec/features/user_signin_spec.rb
@@ -21,11 +21,4 @@ feature 'User sign in' do
     expect(page).to have_content "Welcome, #{user.nom_name}"
   end
 
-
-  def sign_in(email:, password:)
-    visit '/sessions/new'
-    fill_in :email, with: email
-    fill_in :password, with: password
-    click_button 'Sign in'
-  end
 end
